MainApp/requirements.txt
fastapi>=0.110.1
pydantic==2.3.0
uvicorn>=0.15.0,<0.16.0
SQLAlchemy~=2.0.27
starlette~=0.37.2
python-dotenv~=1.0.1
alembic~=1.13.1
jinja2==3.1.3
aiofiles==0.7.0
asyncpg==0.28.0
pydantic-settings==2.0.3
psycopg==3.1.12
psycopg-binary==3.1.12
passlib~=1.7.4
python-jose==3.3.0
weasyprint~=62.1
httpx~=0.27.0
pytest~=8.2.0
